chore: move GetApps to core package
