+++ core/domain/src/main/kotlin/dev/mslalith/focuslauncher/core/domain/apps/GetAppDrawerAppsWithIconsUseCase.kt
@@ -2,6 +2,7 @@ package dev.mslalith.focuslauncher.core.domain.apps
 
 import dev.mslalith.focuslauncher.core.data.repository.FavoritesRepo
 import dev.mslalith.focuslauncher.core.domain.appdrawer.GetAppDrawerAppsUseCase
+import dev.mslalith.focuslauncher.core.domain.apps.core.GetAppsIconPackAwareUseCase
 import dev.mslalith.focuslauncher.core.model.app.AppWithIconFavorite
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
