+++ core/domain/src/main/kotlin/dev/mslalith/focuslauncher/core/domain/apps/GetFavoriteAppsWithColorUseCase.kt
@@ -1,6 +1,7 @@
 package dev.mslalith.focuslauncher.core.domain.apps
 
 import dev.mslalith.focuslauncher.core.data.repository.FavoritesRepo
+import dev.mslalith.focuslauncher.core.domain.apps.core.GetAppsIconPackAwareUseCase
 import dev.mslalith.focuslauncher.core.model.app.AppWithColor
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
